automation/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
colorama>=0.4.6
